THIS IS MY FIRST COMMIT TO THE SEINFELD REPO
This is actually the second commit, because create-react-app makes
the initial commit.

In the previous commit, I ran the command:
npx create-react-app seinfeld --template typescript.

In this commit, all that I altered was the README.md file, with inital
information about this project. Otherwise, nothing was altered that was
generated by create-react-app.